pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "solo"
version = "0.1.1"
authors = [{ name = "SoloKeys", email = "hello@solokeys.com" }]
description = "Test package"
readme = "README.md"
license = { text = "MIT License" }
requires-python = ">=3.11"
classifiers=[
    "License :: OSI Approved :: MIT License",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
  "click >= 7.1",
  "cryptography",
  "ecdsa",
  "fido2 >= 0.9.1",
  "intelhex",
  "pyserial",
  "pyusb",
  "requests",
]


[project.scripts]
solo = "solo.cli:solo_cli"

[project.urls]
"Homepage" = "https://github.com/solokeys/solo1-cli"
"Bug Tracker" = "https://github.com/solokeys/solo1-cli/issues"

